requirements.txt
pandas
numpy
biopython
crossref-commons
pygraphviz
ipykernel
nbformat
nbclient
jupyter-cache
matplotlib
seaborn
hvplot
flask
Flask-Caching
python-dotenv
click